msm: adsprpc: API to get DSP capability
This API will be initiated from a user application. If the DSP information
is not on the kernel it will make a call to the DSP and retrieve the
information.
Then the information will be cached on the kernel and returned to the user
program. Subsequent API calls will retrieve the previously cached
information from the kernel.

Change-Id: Ib89ad25e857f02f1e1ca02180c0363e1543503b4
Acked-by: Edgar Flores <edgarf@qti.qualcomm.com>
Signed-off-by: Tharun Kumar Merugu <mtharu@codeaurora.org>
Signed-off-by: Radu Vasilascu <glx33807@gmail.com>